test(security): stabilize credential filter coverage
